Move worktree creation logic out of the agent panel (#54200)
Opening this up a separate PR for ease of review, contribution, and
focus. It's pointing at the PR in which we move the worktree picker into
the title bar.

Release Notes:

- N/A
