+++ crates/agent_ui/src/agent_panel.rs
@@ -1,5 +1,5 @@
 use std::{
-    path::{Path, PathBuf},
+    path::PathBuf,
     rc::Rc,
     sync::{
         Arc,
@@ -32,10 +32,10 @@ use zed_actions::{
 use crate::DEFAULT_THREAD_TITLE;
 use crate::thread_metadata_store::{ThreadId, ThreadMetadataStore};
 use crate::{
-    AddContextServer, AgentDiffPane, ConversationView, CopyThreadToClipboard, CreateWorktree,
-    Follow, InlineAssistant, LoadThreadFromClipboard, NewThread, NewWorktreeBranchTarget,
-    OpenActiveThreadAsMarkdown, OpenAgentDiff, OpenHistory, ResetTrialEndUpsell, ResetTrialUpsell,
-    SwitchWorktree, ToggleNavigationMenu, ToggleNewThreadMenu, ToggleOptionsMenu,
+    AddContextServer, AgentDiffPane, ConversationView, CopyThreadToClipboard, Follow,
+    InlineAssistant, LoadThreadFromClipboard, NewThread, OpenActiveThreadAsMarkdown, OpenAgentDiff,
+    OpenHistory, ResetTrialEndUpsell, ResetTrialUpsell, ToggleNavigationMenu, ToggleNewThreadMenu,
+    ToggleOptionsMenu,
     agent_configuration::{AgentConfiguration, AssistantConfigurationEvent},
     conversation_view::{AcpThreadViewEvent, ThreadView},
     ui::EndTrialUpsell,
@@ -49,7 +49,7 @@ use crate::{ManageProfiles, ThreadHistoryViewEvent};
 use crate::{ThreadHistory, agent_connection_store::AgentConnectionStore};
 use agent_settings::AgentSettings;
 use ai_onboarding::AgentPanelOnboarding;
-use anyhow::{Context as _, Result, anyhow};
+use anyhow::{Context as _, Result};
 use client::UserStore;
 use cloud_api_types::Plan;
 use collections::HashMap;
@@ -59,18 +59,14 @@ use extension_host::ExtensionStore;
 use fs::Fs;
 use gpui::{
     Action, Animation, AnimationExt, AnyElement, App, AsyncWindowContext, ClipboardItem, Corner,
-    DismissEvent, Entity, EntityId, EventEmitter, ExternalPaths, FocusHandle, Focusable,
-    KeyContext, Pixels, Subscription, Task, UpdateGlobal, WeakEntity, prelude::*,
-    pulsating_between,
+    DismissEvent, Entity, EventEmitter, ExternalPaths, FocusHandle, Focusable, KeyContext, Pixels,
+    Subscription, Task, UpdateGlobal, WeakEntity, prelude::*, pulsating_between,
 };
 use language::LanguageRegistry;
 use language_model::LanguageModelRegistry;
-use project::project_settings::ProjectSettings;
-use project::trusted_worktrees::{PathTrust, TrustedWorktrees};
 use project::{Project, ProjectPath, Worktree};
 use prompt_store::{PromptStore, UserPromptId};
 use release_channel::ReleaseChannel;
-use remote::RemoteConnectionOptions;
 use rules_library::{RulesLibrary, open_rules_library};
 use settings::TerminalDockPosition;
 use settings::{Settings, update_settings_file};
@@ -83,8 +79,8 @@ use ui::{
 };
 use util::{ResultExt as _, debug_panic};
 use workspace::{
-    CollaboratorId, DockStructure, DraggedSelection, DraggedTab, OpenMode, PathList,
-    SerializedPathList, ToggleWorkspaceSidebar, ToggleZoom, Workspace, WorkspaceId,
+    CollaboratorId, DraggedSelection, DraggedTab, PathList, SerializedPathList,
+    ToggleWorkspaceSidebar, ToggleZoom, Workspace, WorkspaceId,
     dock::{DockPosition, Panel, PanelEvent},
 };
 
@@ -479,28 +475,6 @@ pub fn init(cx: &mut App) {
                             });
                         });
                     },
-                )
-                .register_action(
-                    |workspace: &mut Workspace, action: &CreateWorktree, window, cx| {
-                        let previous_state =
-                            AgentPanel::capture_workspace_state(workspace, window, cx);
-                        if let Some(panel) = workspace.panel::<AgentPanel>(cx) {
-                            panel.update(cx, |panel, cx| {
-                                panel.create_worktree(action, previous_state, window, cx);
-                            });
-                        }
-                    },
-                )
-                .register_action(
-                    |workspace: &mut Workspace, action: &SwitchWorktree, window, cx| {
-                        let previous_state =
-                            AgentPanel::capture_workspace_state(workspace, window, cx);
-                        if let Some(panel) = workspace.panel::<AgentPanel>(cx) {
-                            panel.update(cx, |panel, cx| {
-                                panel.switch_to_worktree(action, previous_state, window, cx);
-                            });
-                        }
-                    },
                 );
         },
     )
@@ -650,61 +624,6 @@ enum WhichFontSize {
     None,
 }
 
-#[derive(Clone, Debug)]
-pub enum WorktreeCreationStatus {
-    Creating(SharedString),
-    Loading(SharedString),
-    Error(SharedString),
-}
-
-#[derive(Clone, Debug)]
-enum WorktreeCreationArgs {
-    New {
-        worktree_name: Option<String>,
-        branch_target: NewWorktreeBranchTarget,
-    },
-    Linked {
-        worktree_path: PathBuf,
-        display_name: String,
-    },
-}
-
-struct PreviousWorkspaceState {
-    dock_structure: DockStructure,
-    open_file_paths: Vec<PathBuf>,
-    active_file_path: Option<PathBuf>,
-}
-
-#[cfg(test)]
-impl PreviousWorkspaceState {
-    /// An empty state with all docks hidden and no open files.
-    fn empty() -> Self {
-        use workspace::DockData;
-
-        Self {
-            dock_structure: DockStructure {
-                left: DockData {
-                    visible: false,
-                    active_panel: None,
-                    zoom: false,
-                },
-                right: DockData {
-                    visible: false,
-                    active_panel: None,
-                    zoom: false,
-                },
-                bottom: DockData {
-                    visible: false,
-                    active_panel: None,
-                    zoom: false,
-                },
-            },
-            open_file_paths: Vec::new(),
-            active_file_path: None,
-        }
-    }
-}
-
 impl BaseView {
     pub fn which_font_size_used(&self) -> WhichFontSize {
         WhichFontSize::AgentFont
@@ -750,11 +669,11 @@ pub struct AgentPanel {
     new_user_onboarding: Entity<AgentPanelOnboarding>,
     new_user_onboarding_upsell_dismissed: AtomicBool,
     selected_agent: Agent,
-    worktree_creation_status: Option<(EntityId, WorktreeCreationStatus)>,
     _thread_view_subscription: Option<Subscription>,
     _active_thread_focus_subscription: Option<Subscription>,
-    _worktree_creation_task: Option<Task<()>>,
     show_trust_workspace_message: bool,
+    did_stash_worktree_draft: bool,
+    pending_worktree_draft: Option<String>,
     _base_view_observation: Option<Subscription>,
     _draft_editor_observation: Option<Subscription>,
 }
@@ -1116,15 +1035,75 @@ impl AgentPanel {
             new_user_onboarding: onboarding,
             thread_store,
             selected_agent: Agent::default(),
-            worktree_creation_status: None,
             _thread_view_subscription: None,
             _active_thread_focus_subscription: None,
-            _worktree_creation_task: None,
             show_trust_workspace_message: false,
+            did_stash_worktree_draft: false,
+            pending_worktree_draft: None,
             new_user_onboarding_upsell_dismissed: AtomicBool::new(OnboardingUpsell::dismissed(cx)),
             _base_view_observation: None,
             _draft_editor_observation: None,
         };
+
+        // Subscribe to the workspace to detect worktree creation changes.
+        // When creation starts (label becomes Some): this is the SOURCE panel.
+        //   → Stash draft text via the workspace helper and set a flag.
+        // When creation ends (label becomes None): this is the DESTINATION panel
+        //   (the flag is false on fresh panels) → pick up the stashed text.
+        if let Some(workspace_entity) = panel.workspace.upgrade() {
+            cx.subscribe_in(
+                &workspace_entity,
+                window,
+                |this, workspace, event: &workspace::Event, window, cx| {
+                    if !matches!(event, workspace::Event::WorktreeCreationChanged) {
+                        return;
+                    }
+
+                    let is_creating = workspace
+                        .read(cx)
+                        .active_worktree_creation()
+                        .label
+                        .is_some();
+
+                    if is_creating {
+                        let draft_prompt = this.active_thread_view(cx).and_then(|thread_view| {
+                            let text = thread_view.read(cx).message_editor.read(cx).text(cx);
+                            if text.is_empty() { None } else { Some(text) }
+                        });
+
+                        let multi_workspace = workspace.read(cx).multi_workspace().cloned();
+                        if let Some(multi_workspace) = multi_workspace.and_then(|mw| mw.upgrade()) {
+                            multi_workspace.update(cx, |mw, _cx| {
+                                mw.pending_worktree_switch_text = draft_prompt;
+                            });
+                        }
+
+                        this.did_stash_worktree_draft = true;
+                    } else if !this.did_stash_worktree_draft {
+                        let multi_workspace = workspace.read(cx).multi_workspace().cloned();
+                        let pending_text =
+                            multi_workspace.and_then(|mw| mw.upgrade()).and_then(|mw| {
+                                mw.update(cx, |mw, _cx| mw.pending_worktree_switch_text.take())
+                            });
+
+                        if let Some(text) = pending_text {
+                            if let Some(thread_view) = this.active_thread_view(cx) {
+                                thread_view.update(cx, |thread_view, cx| {
+                                    thread_view.message_editor.update(cx, |editor, cx| {
+                                        editor.clear(window, cx);
+                                        editor.insert_text(&text, window, cx);
+                                    });
+                                });
+                            } else {
+                                this.pending_worktree_draft = Some(text);
+                                this.ensure_thread_initialized(window, cx);
+                            }
+                        }
+                    }
+                },
+            )
+            .detach();
+        }
 
         // Initial sync of agent servers from extensions
         panel.sync_agent_servers_from_extensions(cx);
@@ -2262,6 +2241,7 @@ impl AgentPanel {
                     |this, server_view, window, cx| {
                         this._thread_view_subscription =
                             Self::subscribe_to_active_thread_view(&server_view, window, cx);
+                        this.apply_pending_worktree_draft(window, cx);
                         cx.emit(AgentPanelEvent::ActiveViewChanged);
                         this.serialize(cx);
                         cx.notify();
@@ -2659,986 +2639,6 @@ impl AgentPanel {
                 .is_some_and(|active| active.entity_id() == draft.entity_id())
         })
     }
-
-    // TODO: The mapping from workspace root paths to git repositories needs a
-    // unified approach across the codebase: this method, `sidebar::is_root_repo`,
-    // thread persistence (which PathList is saved to the database), and thread
-    // querying (which PathList is used to read threads back). All of these need
-    // to agree on how repos are resolved for a given workspace, especially in
-    // multi-root and nested-repo configurations.
-    /// Partitions the project's visible worktrees into git-backed repositories
-    /// and plain (non-git) paths. Git repos will have worktrees created for
-    /// them; non-git paths are carried over to the new workspace as-is.
-    ///
-    /// When multiple worktrees map to the same repository, the most specific
-    /// match wins (deepest work directory path), with a deterministic
-    /// tie-break on entity id. Each repository appears at most once.
-    fn classify_worktrees(
-        &self,
-        cx: &App,
-    ) -> (Vec<Entity<project::git_store::Repository>>, Vec<PathBuf>) {
-        let project = &self.project;
-        let repositories = project.read(cx).repositories(cx).clone();
-        let mut git_repos: Vec<Entity<project::git_store::Repository>> = Vec::new();
-        let mut non_git_paths: Vec<PathBuf> = Vec::new();
-        let mut seen_repo_ids = std::collections::HashSet::new();
-
-        for worktree in project.read(cx).visible_worktrees(cx) {
-            let wt_path = worktree.read(cx).abs_path();
-
-            let matching_repo = repositories
-                .iter()
-                .filter_map(|(id, repo)| {
-                    let work_dir = repo.read(cx).work_directory_abs_path.clone();
-                    if wt_path.starts_with(work_dir.as_ref()) {
-                        Some((*id, repo.clone(), work_dir.as_ref().components().count()))
-                    } else {
-                        None
-                    }
-                })
-                .max_by(
-                    |(left_id, _left_repo, left_depth), (right_id, _right_repo, right_depth)| {
-                        left_depth
-                            .cmp(right_depth)
-                            .then_with(|| left_id.cmp(right_id))
-                    },
-                );
-
-            if let Some((id, repo, _)) = matching_repo {
-                if seen_repo_ids.insert(id) {
-                    git_repos.push(repo);
-                }
-            } else {
-                non_git_paths.push(wt_path.to_path_buf());
-            }
-        }
-
-        (git_repos, non_git_paths)
-    }
-
-    fn resolve_worktree_branch_target(
-        branch_target: &NewWorktreeBranchTarget,
-    ) -> (Option<String>, Option<String>) {
-        match branch_target {
-            NewWorktreeBranchTarget::CurrentBranch => (None, None),
-            NewWorktreeBranchTarget::ExistingBranch { name } => {
-                (Some(name.clone()), Some(name.clone()))
-            }
-            NewWorktreeBranchTarget::CreateBranch { name, from_ref } => {
-                (Some(name.clone()), from_ref.clone())
-            }
-        }
-    }
-
-    fn maybe_propagate_worktree_trust(
-        this: &WeakEntity<Self>,
-        new_workspace: &Entity<workspace::Workspace>,
-        paths: &[PathBuf],
-        cx: &mut AsyncWindowContext,
-    ) {
-        cx.update(|_, cx| {
-            if ProjectSettings::get_global(cx).session.trust_all_worktrees {
-                return;
-            }
-            let Some(trusted_store) = TrustedWorktrees::try_get_global(cx) else {
-                return;
-            };
-
-            let source_is_trusted = this
-                .upgrade()
-                .map(|panel| {
-                    let source_worktree_store = panel.read(cx).project.read(cx).worktree_store();
-                    !trusted_store
-                        .read(cx)
-                        .has_restricted_worktrees(&source_worktree_store, cx)
-                })
-                .unwrap_or(false);
-
-            if !source_is_trusted {
-                return;
-            }
-
-            let worktree_store = new_workspace.read(cx).project().read(cx).worktree_store();
-            let paths_to_trust: HashSet<_> = paths
-                .iter()
-                .filter_map(|path| {
-                    let (worktree, _) = worktree_store.read(cx).find_worktree(path, cx)?;
-                    Some(PathTrust::Worktree(worktree.read(cx).id()))
-                })
-                .collect();
-
-            if !paths_to_trust.is_empty() {
-                trusted_store.update(cx, |store, cx| {
-                    store.trust(&worktree_store, paths_to_trust, cx);
-                });
-            }
-        })
-        .ok();
-    }
-
-    /// Kicks off an async git-worktree creation for each repository. Returns:
-    ///
-    /// - `creation_infos`: a vec of `(repo, new_path, receiver)` tuples—the
-    ///   receiver resolves once the git worktree command finishes.
-    /// - `path_remapping`: `(old_work_dir, new_worktree_path)` pairs used
-    ///   later to remap open editor tabs into the new workspace.
-    fn start_worktree_creations(
-        git_repos: &[Entity<project::git_store::Repository>],
-        worktree_name: Option<String>,
-        existing_worktree_names: &[String],
-        existing_worktree_paths: &HashSet<PathBuf>,
-        base_ref: Option<String>,
-        worktree_directory_setting: &str,
-        rng: &mut impl rand::Rng,
-        cx: &mut Context<Self>,
-    ) -> Result<(
-        Vec<(
-            Entity<project::git_store::Repository>,
-            PathBuf,
-            futures::channel::oneshot::Receiver<Result<()>>,
-        )>,
-        Vec<(PathBuf, PathBuf)>,
-    )> {
-        let mut creation_infos = Vec::new();
-        let mut path_remapping = Vec::new();
-
-        let worktree_name = worktree_name.unwrap_or_else(|| {
-            let existing_refs: Vec<&str> =
-                existing_worktree_names.iter().map(|s| s.as_str()).collect();
-            crate::worktree_names::generate_worktree_name(&existing_refs, rng)
-                .unwrap_or_else(|| "worktree".to_string())
-        });
-
-        for repo in git_repos {
-            let (work_dir, new_path, receiver) = repo.update(cx, |repo, _cx| {
-                let new_path =
-                    repo.path_for_new_linked_worktree(&worktree_name, worktree_directory_setting)?;
-                if existing_worktree_paths.contains(&new_path) {
-                    anyhow::bail!("A worktree already exists at {}", new_path.display());
-                }
-                let target = git::repository::CreateWorktreeTarget::Detached {
-                    base_sha: base_ref.clone(),
-                };
-                let receiver = repo.create_worktree(target, new_path.clone());
-                let work_dir = repo.work_directory_abs_path.clone();
-                anyhow::Ok((work_dir, new_path, receiver))
-            })?;
-            path_remapping.push((work_dir.to_path_buf(), new_path.clone()));
-            creation_infos.push((repo.clone(), new_path, receiver));
-        }
-
-        Ok((creation_infos, path_remapping))
-    }
-
-    /// Waits for every in-flight worktree creation to complete. If any
-    /// creation fails, all successfully-created worktrees are rolled back
-    /// (removed) so the project isn't left in a half-migrated state.
-    async fn await_and_rollback_on_failure(
-        creation_infos: Vec<(
-            Entity<project::git_store::Repository>,
-            PathBuf,
-            futures::channel::oneshot::Receiver<Result<()>>,
-        )>,
-        fs: Arc<dyn Fs>,
-        cx: &mut AsyncWindowContext,
-    ) -> Result<Vec<PathBuf>> {
-        let mut created_paths: Vec<PathBuf> = Vec::new();
-        let mut repos_and_paths: Vec<(Entity<project::git_store::Repository>, PathBuf)> =
-            Vec::new();
-        let mut first_error: Option<anyhow::Error> = None;
-
-        for (repo, new_path, receiver) in creation_infos {
-            repos_and_paths.push((repo.clone(), new_path.clone()));
-            match receiver.await {
-                Ok(Ok(())) => {
-                    created_paths.push(new_path);
-                }
-                Ok(Err(err)) => {
-                    if first_error.is_none() {
-                        first_error = Some(err);
-                    }
-                }
-                Err(_canceled) => {
-                    if first_error.is_none() {
-                        first_error = Some(anyhow!("Worktree creation was canceled"));
-                    }
-                }
-            }
-        }
-
-        let Some(err) = first_error else {
-            return Ok(created_paths);
-        };
-
-        // Rollback all attempted worktrees (both successful and failed,
-        // since a failed creation may have left an orphan directory).
-        let mut rollback_futures = Vec::new();
-        for (rollback_repo, rollback_path) in &repos_and_paths {
-            let receiver = cx
-                .update(|_, cx| {
-                    rollback_repo.update(cx, |repo, _cx| {
-                        repo.remove_worktree(rollback_path.clone(), true)
-                    })
-                })
-                .ok();
-
-            rollback_futures.push((rollback_path.clone(), receiver));
-        }
-
-        let mut rollback_failures: Vec<String> = Vec::new();
-        for (path, receiver_opt) in rollback_futures {
-            let mut git_remove_failed = false;
-
-            if let Some(receiver) = receiver_opt {
-                match receiver.await {
-                    Ok(Ok(())) => {}
-                    Ok(Err(rollback_err)) => {
-                        log::error!(
-                            "git worktree remove failed for {}: {rollback_err}",
-                            path.display()
-                        );
-                        git_remove_failed = true;
-                    }
-                    Err(canceled) => {
-                        log::error!(
-                            "git worktree remove failed for {}: {canceled}",
-                            path.display()
-                        );
-                        git_remove_failed = true;
-                    }
-                }
-            } else {
-                log::error!(
-                    "failed to dispatch git worktree remove for {}",
-                    path.display()
-                );
-                git_remove_failed = true;
-            }
-
-            // `git worktree remove` normally removes this directory, but since
-            // `git worktree remove` failed (or wasn't dispatched), manually rm the directory.
-            if git_remove_failed {
-                if let Err(fs_err) = fs
-                    .remove_dir(
-                        &path,
-                        fs::RemoveOptions {
-                            recursive: true,
-                            ignore_if_not_exists: true,
-                        },
-                    )
-                    .await
-                {
-                    let msg = format!("{}: failed to remove directory: {fs_err}", path.display());
-                    log::error!("{}", msg);
-                    rollback_failures.push(msg);
-                }
-            }
-        }
-        let mut error_message = format!("Failed to create worktree: {err}");
-        if !rollback_failures.is_empty() {
-            error_message.push_str("\n\nFailed to clean up: ");
-            error_message.push_str(&rollback_failures.join(", "));
-        }
-        Err(anyhow!(error_message))
-    }
-
-    /// Attempts to check out a branch in a newly created worktree.
-    /// First tries checking out an existing branch, then tries creating a new
-    /// branch. If both fail, the worktree stays in detached HEAD state.
-    async fn try_checkout_branch_in_worktree(
-        repo: &Entity<project::git_store::Repository>,
-        branch_name: &str,
-        worktree_path: &Path,
-        cx: &mut AsyncWindowContext,
-    ) {
-        // First, try checking out the branch (it may already exist).
-        let Ok(receiver) = cx.update(|_, cx| {
-            repo.update(cx, |repo, _cx| {
-                repo.checkout_branch_in_worktree(
-                    branch_name.to_string(),
-                    worktree_path.to_path_buf(),
-                    false,
-                )
-            })
-        }) else {
-            log::warn!(
-                "Failed to check out branch {branch_name} for worktree at {}. \
-                 Staying in detached HEAD state.",
-                worktree_path.display(),
-            );
-
-            return;
-        };
-
-        let Ok(result) = receiver.await else {
-            log::warn!(
-                "Branch checkout was canceled for worktree at {}. \
-                 Staying in detached HEAD state.",
-                worktree_path.display()
-            );
-
-            return;
-        };
-
-        if let Err(err) = result {
-            log::info!(
-                "Failed to check out branch '{branch_name}' in worktree at {}, \
-                         will try creating it: {err}",
-                worktree_path.display()
-            );
-        } else {
-            log::info!(
-                "Checked out branch '{branch_name}' in worktree at {}",
-                worktree_path.display()
-            );
-
-            return;
-        }
-
-        // Checkout failed, so try creating the branch.
-        let create_result = cx.update(|_, cx| {
-            repo.update(cx, |repo, _cx| {
-                repo.checkout_branch_in_worktree(
-                    branch_name.to_string(),
-                    worktree_path.to_path_buf(),
-                    true,
-                )
-            })
-        });
-
-        match create_result {
-            Ok(receiver) => match receiver.await {
-                Ok(Ok(())) => {
-                    log::info!(
-                        "Created and checked out branch '{branch_name}' in worktree at {}",
-                        worktree_path.display()
-                    );
-                }
-                Ok(Err(err)) => {
-                    log::warn!(
-                        "Failed to create branch '{branch_name}' in worktree at {}: {err}. \
-                         Staying in detached HEAD state.",
-                        worktree_path.display()
-                    );
-                }
-                Err(_) => {
-                    log::warn!(
-                        "Branch creation was canceled for worktree at {}. \
-                         Staying in detached HEAD state.",
-                        worktree_path.display()
-                    );
-                }
-            },
-            Err(err) => {
-                log::warn!(
-                    "Failed to dispatch branch creation for worktree at {}: {err}. \
-                     Staying in detached HEAD state.",
-                    worktree_path.display(),
-                );
-            }
-        }
-    }
-
-    fn capture_workspace_state(
-        workspace: &Workspace,
-        window: &Window,
-        cx: &App,
-    ) -> PreviousWorkspaceState {
-        let dock_structure = workspace.capture_dock_state(window, cx);
-        let open_file_paths = workspace.open_item_abs_paths(cx);
-        let active_file_path = workspace
-            .active_item(cx)
-            .and_then(|item| item.project_path(cx))
-            .and_then(|pp| workspace.project().read(cx).absolute_path(&pp, cx));
-
-        PreviousWorkspaceState {
-            dock_structure,
-            open_file_paths,
-            active_file_path,
-        }
-    }
-
-    fn create_worktree(
-        &mut self,
-        action: &CreateWorktree,
-        previous_workspace_state: PreviousWorkspaceState,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if !self.project_has_git_repository(cx) {
-            log::error!("create_worktree: no git repository in the project");
-            return;
-        }
-        if self.project.read(cx).is_via_collab() {
-            log::error!("create_worktree: not supported in collab projects");
-            return;
-        }
-        if matches!(
-            self.worktree_creation_status,
-            Some((
-                _,
-                WorktreeCreationStatus::Creating(_) | WorktreeCreationStatus::Loading(_)
-            ))
-        ) {
-            return;
-        }
-
-        let content = self.take_active_initial_content(cx);
-        let content_blocks = match content {
-            Some(AgentInitialContent::ContentBlock { blocks, .. }) => blocks,
-            _ => Vec::new(),
-        };
-
-        self.handle_worktree_requested(
-            content_blocks,
-            WorktreeCreationArgs::New {
-                worktree_name: action.worktree_name.clone(),
-                branch_target: action.branch_target.clone(),
-            },
-            previous_workspace_state,
-            window,
-            cx,
-        );
-    }
-
-    fn switch_to_worktree(
-        &mut self,
-        action: &SwitchWorktree,
-        previous_workspace_state: PreviousWorkspaceState,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if !self.project_has_git_repository(cx) {
-            log::error!("switch_to_worktree: no git repository in the project");
-            return;
-        }
-        if self.project.read(cx).is_via_collab() {
-            log::error!("switch_to_worktree: not supported in collab projects");
-            return;
-        }
-        if matches!(
-            self.worktree_creation_status,
-            Some((
-                _,
-                WorktreeCreationStatus::Creating(_) | WorktreeCreationStatus::Loading(_)
-            ))
-        ) {
-            return;
-        }
-
-        let content = self.take_active_initial_content(cx);
-        let content_blocks = match content {
-            Some(AgentInitialContent::ContentBlock { blocks, .. }) => blocks,
-            _ => Vec::new(),
-        };
-
-        self.handle_worktree_requested(
-            content_blocks,
-            WorktreeCreationArgs::Linked {
-                worktree_path: action.path.clone(),
-                display_name: action.display_name.clone(),
-            },
-            previous_workspace_state,
-            window,
-            cx,
-        );
-    }
-
-    fn set_worktree_creation_error(
-        &mut self,
-        message: SharedString,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if let Some((_, status)) = &mut self.worktree_creation_status {
-            *status = WorktreeCreationStatus::Error(message);
-        }
-        if matches!(self.base_view, BaseView::Uninitialized) {
-            let selected_agent = self.selected_agent.clone();
-            self.new_agent_thread(selected_agent, window, cx);
-        }
-        cx.notify();
-    }
-
-    fn handle_worktree_requested(
-        &mut self,
-        content: Vec<acp::ContentBlock>,
-        args: WorktreeCreationArgs,
-        previous_workspace_state: PreviousWorkspaceState,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if matches!(
-            self.worktree_creation_status,
-            Some((
-                _,
-                WorktreeCreationStatus::Creating(_) | WorktreeCreationStatus::Loading(_)
-            ))
-        ) {
-            return;
-        }
-
-        let conversation_view_id = self
-            .active_conversation_view()
-            .map(|v| v.entity_id())
-            .unwrap_or_else(|| EntityId::from(0u64));
-        let display_name: SharedString = match &args {
-            WorktreeCreationArgs::New {
-                worktree_name: Some(name),
-                ..
-            } => name.clone().into(),
-            WorktreeCreationArgs::New { .. } => "worktree".into(),
-            WorktreeCreationArgs::Linked { display_name, .. } => display_name.clone().into(),
-        };
-        let global_label = display_name.clone();
-        let status = if matches!(args, WorktreeCreationArgs::Linked { .. }) {
-            WorktreeCreationStatus::Loading(display_name)
-        } else {
-            WorktreeCreationStatus::Creating(display_name)
-        };
-        self.worktree_creation_status = Some((conversation_view_id, status));
-        cx.set_global(workspace::ActiveWorktreeCreation {
-            label: Some(global_label),
-        });
-        cx.notify();
-
-        let (git_repos, non_git_paths) = self.classify_worktrees(cx);
-
-        if matches!(args, WorktreeCreationArgs::New { .. }) && git_repos.is_empty() {
-            self.set_worktree_creation_error(
-                "No git repositories found in the project".into(),
-                window,
-                cx,
-            );
-            return;
-        }
-
-        let remote_connection_options = self.project.read(cx).remote_connection_options(cx);
-
-        if remote_connection_options.is_some() {
-            let is_disconnected = self
-                .project
-                .read(cx)
-                .remote_client()
-                .is_some_and(|client| client.read(cx).is_disconnected());
-            if is_disconnected {
-                self.set_worktree_creation_error(
-                    "Cannot create worktree: remote connection is not active".into(),
-                    window,
-                    cx,
-                );
-                return;
-            }
-        }
-
-        let workspace = self.workspace.clone();
-        let window_handle = window
-            .window_handle()
-            .downcast::<workspace::MultiWorkspace>();
-
-        let selected_agent = self.selected_agent();
-
-        let git_repo_work_dirs: Vec<PathBuf> = git_repos
-            .iter()
-            .map(|repo| repo.read(cx).work_directory_abs_path.to_path_buf())
-            .collect();
-
-        let task = cx.spawn_in(window, async move |this, cx| {
-            let (all_paths, path_remapping, has_non_git) = match args {
-                WorktreeCreationArgs::New {
-                    worktree_name,
-                    branch_target,
-                } => {
-                    let worktree_receivers: Vec<_> = this.update_in(cx, |_this, _window, cx| {
-                        git_repos
-                            .iter()
-                            .map(|repo| repo.update(cx, |repo, _cx| repo.worktrees()))
-                            .collect()
-                    })?;
-                    let worktree_directory_setting = this.update_in(cx, |_this, _window, cx| {
-                        ProjectSettings::get_global(cx)
-                            .git
-                            .worktree_directory
-                            .clone()
-                    })?;
-
-                    let mut existing_worktree_names = Vec::new();
-                    let mut existing_worktree_paths = HashSet::default();
-                    for result in futures::future::join_all(worktree_receivers).await {
-                        match result {
-                            Ok(Ok(worktrees)) => {
-                                for worktree in worktrees {
-                                    if let Some(name) = worktree
-                                        .path
-                                        .parent()
-                                        .and_then(|p| p.file_name())
-                                        .and_then(|n| n.to_str())
-                                    {
-                                        existing_worktree_names.push(name.to_string());
-                                    }
-                                    existing_worktree_paths.insert(worktree.path.clone());
-                                }
-                            }
-                            Ok(Err(err)) => {
-                                Err::<(), _>(err).log_err();
-                            }
-                            Err(_) => {}
-                        }
-                    }
-
-                    let mut rng = rand::rng();
-
-                    let (branch_to_checkout, base_ref) =
-                        Self::resolve_worktree_branch_target(&branch_target);
-
-                    let (creation_infos, path_remapping) =
-                        match this.update_in(cx, |_this, _window, cx| {
-                            Self::start_worktree_creations(
-                                &git_repos,
-                                worktree_name,
-                                &existing_worktree_names,
-                                &existing_worktree_paths,
-                                base_ref,
-                                &worktree_directory_setting,
-                                &mut rng,
-                                cx,
-                            )
-                        }) {
-                            Ok(Ok(result)) => result,
-                            Ok(Err(err)) | Err(err) => {
-                                this.update_in(cx, |this, window, cx| {
-                                    this.set_worktree_creation_error(
-                                        format!("Failed to validate worktree directory: {err}")
-                                            .into(),
-                                        window,
-                                        cx,
-                                    );
-                                })
-                                .log_err();
-                                return anyhow::Ok(());
-                            }
-                        };
-
-                    let repo_paths: Vec<(Entity<project::git_store::Repository>, PathBuf)> =
-                        creation_infos
-                            .iter()
-                            .map(|(repo, path, _)| (repo.clone(), path.clone()))
-                            .collect();
-
-                    let fs = cx.update(|_, cx| <dyn Fs>::global(cx))?;
-
-                    let created_paths =
-                        match Self::await_and_rollback_on_failure(creation_infos, fs, cx).await {
-                            Ok(paths) => paths,
-                            Err(err) => {
-                                this.update_in(cx, |this, window, cx| {
-                                    this.set_worktree_creation_error(
-                                        format!("{err}").into(),
-                                        window,
-                                        cx,
-                                    );
-                                })?;
-                                return anyhow::Ok(());
-                            }
-                        };
-
-                    if let Some(ref branch_name) = branch_to_checkout {
-                        for (repo, worktree_path) in &repo_paths {
-                            Self::try_checkout_branch_in_worktree(
-                                repo,
-                                branch_name,
-                                worktree_path,
-                                cx,
-                            )
-                            .await;
-                        }
-                    }
-
-                    let mut all_paths = created_paths;
-                    let has_non_git = !non_git_paths.is_empty();
-                    all_paths.extend(non_git_paths.iter().cloned());
-                    (all_paths, path_remapping, has_non_git)
-                }
-                WorktreeCreationArgs::Linked { worktree_path, .. } => {
-                    let path_remapping: Vec<(PathBuf, PathBuf)> = git_repo_work_dirs
-                        .iter()
-                        .map(|work_dir| (work_dir.clone(), worktree_path.clone()))
-                        .collect();
-                    let mut all_paths = vec![worktree_path];
-                    let has_non_git = !non_git_paths.is_empty();
-                    all_paths.extend(non_git_paths.iter().cloned());
-                    (all_paths, path_remapping, has_non_git)
-                }
-            };
-
-            if workspace.upgrade().is_none() {
-                this.update_in(cx, |this, window, cx| {
-                    this.set_worktree_creation_error(
-                        "Workspace no longer available".into(),
-                        window,
-                        cx,
-                    );
-                })?;
-                return anyhow::Ok(());
-            }
-
-            let this_for_error = this.clone();
-            if let Err(err) = Self::open_worktree_workspace_and_start_thread(
-                this,
-                all_paths,
-                window_handle,
-                previous_workspace_state,
-                path_remapping,
-                non_git_paths,
-                has_non_git,
-                content,
-                selected_agent,
-                remote_connection_options,
-                cx,
-            )
-            .await
-            {
-                this_for_error
-                    .update_in(cx, |this, window, cx| {
-                        this.set_worktree_creation_error(
-                            format!("Failed to set up workspace: {err}").into(),
-                            window,
-                            cx,
-                        );
-                    })
-                    .log_err();
-            }
-            anyhow::Ok(())
-        });
-
-        self._worktree_creation_task = Some(cx.background_spawn(async move {
-            task.await.log_err();
-        }));
-    }
-
-    async fn open_worktree_workspace_and_start_thread(
-        this: WeakEntity<Self>,
-        all_paths: Vec<PathBuf>,
-        window_handle: Option<gpui::WindowHandle<workspace::MultiWorkspace>>,
-        previous_workspace_state: PreviousWorkspaceState,
-        path_remapping: Vec<(PathBuf, PathBuf)>,
-        non_git_paths: Vec<PathBuf>,
-        has_non_git: bool,
-        content: Vec<acp::ContentBlock>,
-        selected_agent: Option<Agent>,
-        remote_connection_options: Option<RemoteConnectionOptions>,
-        cx: &mut AsyncWindowContext,
-    ) -> Result<()> {
-        let window_handle = window_handle
-            .ok_or_else(|| anyhow!("No window handle available for workspace creation"))?;
-
-        let (workspace_task, modal_workspace) =
-            window_handle.update(cx, |multi_workspace, window, cx| {
-                let path_list = PathList::new(&all_paths);
-                let active_workspace = multi_workspace.workspace().clone();
-                let modal_workspace = active_workspace.clone();
-
-                let dock_structure = previous_workspace_state.dock_structure;
-                let init = Box::new(
-                    move |workspace: &mut Workspace,
-                          window: &mut Window,
-                          cx: &mut Context<Workspace>| {
-                        workspace.set_dock_structure(dock_structure, window, cx);
-                    },
-                );
-
-                let task = multi_workspace.find_or_create_workspace(
-                    path_list,
-                    remote_connection_options,
-                    None,
-                    move |connection_options, window, cx| {
-                        remote_connection::connect_with_modal(
-                            &active_workspace,
-                            connection_options,
-                            window,
-                            cx,
-                        )
-                    },
-                    &[],
-                    Some(init),
-                    OpenMode::Add,
-                    window,
-                    cx,
-                );
-                (task, modal_workspace)
-            })?;
-
-        let result = workspace_task.await;
-        remote_connection::dismiss_connection_modal(&modal_workspace, cx);
-        let new_workspace = result?;
-
-        let panels_task = new_workspace.update(cx, |workspace, _cx| workspace.take_panels_task());
-
-        if let Some(task) = panels_task {
-            task.await.log_err();
-        }
-
-        new_workspace
-            .update(cx, |workspace, cx| {
-                workspace.project().read(cx).wait_for_initial_scan(cx)
-            })
-            .await;
-
-        new_workspace
-            .update(cx, |workspace, cx| {
-                let repos = workspace
-                    .project()
-                    .read(cx)
-                    .repositories(cx)
-                    .values()
-                    .cloned()
-                    .collect::<Vec<_>>();
-
-                let tasks = repos
-                    .into_iter()
-                    .map(|repo| repo.update(cx, |repo, _| repo.barrier()));
-                futures::future::join_all(tasks)
-            })
-            .await;
-
-        Self::maybe_propagate_worktree_trust(&this, &new_workspace, &all_paths, cx);
-
-        let initial_content = AgentInitialContent::ContentBlock {
-            blocks: content,
-            auto_submit: false,
-        };
-
-        window_handle.update(cx, |_multi_workspace, window, cx| {
-            new_workspace.update(cx, |workspace, cx| {
-                if has_non_git {
-                    let toast_id = workspace::notifications::NotificationId::unique::<AgentPanel>();
-                    workspace.show_toast(
-                        workspace::Toast::new(
-                            toast_id,
-                            "Some project folders are not git repositories. \
-                             They were included as-is without creating a worktree.",
-                        ),
-                        cx,
-                    );
-                }
-
-                // Remap every previously-open file path into the new worktree.
-                // Paths that can't be remapped (e.g. files that don't exist on
-                // the target branch) are silently skipped — best-effort.
-                let remap_path = |original_path: PathBuf| -> Option<PathBuf> {
-                    let best_match = path_remapping
-                        .iter()
-                        .filter_map(|(old_root, new_root)| {
-                            original_path.strip_prefix(old_root).ok().map(|relative| {
-                                (old_root.components().count(), new_root.join(relative))
-                            })
-                        })
-                        .max_by_key(|(depth, _)| *depth);
-
-                    if let Some((_, remapped_path)) = best_match {
-                        return Some(remapped_path);
-                    }
-
-                    for non_git in &non_git_paths {
-                        if original_path.starts_with(non_git) {
-                            return Some(original_path);
-                        }
-                    }
-                    None
-                };
-
-                let remapped_active_path = previous_workspace_state
-                    .active_file_path
-                    .and_then(|p| remap_path(p));
-
-                // Collect all remapped paths, deduplicating and preserving order.
-                // The active file is placed last so it ends up as the focused tab.
-                let mut paths_to_open: Vec<PathBuf> = Vec::new();
-                let mut seen = HashSet::default();
-                for path in previous_workspace_state.open_file_paths {
-                    if let Some(remapped) = remap_path(path) {
-                        if remapped_active_path.as_ref() != Some(&remapped)
-                            && seen.insert(remapped.clone())
-                        {
-                            paths_to_open.push(remapped);
-                        }
-                    }
-                }
-
-                if let Some(active) = &remapped_active_path {
-                    if seen.insert(active.clone()) {
-                        paths_to_open.push(active.clone());
-                    }
-                }
-
-                if !paths_to_open.is_empty() {
-                    let open_task = workspace.open_paths(
-                        paths_to_open,
-                        workspace::OpenOptions {
-                            focus: Some(false),
-                            ..Default::default()
-                        },
-                        None,
-                        window,
-                        cx,
-                    );
-                    cx.spawn(async move |_, _| -> anyhow::Result<()> {
-                        for item in open_task.await.into_iter().flatten() {
-                            // Best-effort: files that don't exist on the target
-                            // branch will fail to open and that's fine.
-                            item.log_err();
-                        }
-                        Ok(())
-                    })
-                    .detach_and_log_err(cx);
-                }
-            });
-        })?;
-
-        window_handle.update(cx, |multi_workspace, window, cx| {
-            multi_workspace.activate(new_workspace.clone(), window, cx);
-
-            new_workspace.update(cx, |workspace, cx| {
-                workspace.run_create_worktree_tasks(window, cx);
-
-                workspace.focus_panel::<AgentPanel>(window, cx);
-
-                if let Some(panel) = workspace.panel::<AgentPanel>(cx) {
-                    panel.update(cx, |panel, cx| {
-                        panel.external_thread(
-                            selected_agent,
-                            None,
-                            None,
-                            None,
-                            Some(initial_content),
-                            true,
-                            "agent_panel",
-                            window,
-                            cx,
-                        );
-                    });
-                }
-            })
-        })?;
-
-        this.update_in(cx, |this, window, cx| {
-            this.worktree_creation_status = None;
-            cx.set_global(workspace::ActiveWorktreeCreation { label: None });
-
-            if let Some(thread_view) = this.active_thread_view(cx) {
-                thread_view.update(cx, |thread_view, cx| {
-                    thread_view
-                        .message_editor
-                        .update(cx, |editor, cx| editor.clear(window, cx));
-                });
-            }
-
-            this.serialize(cx);
-            cx.notify();
-        })?;
-
-        anyhow::Ok(())
-    }
 }
 
 impl Focusable for AgentPanel {
@@ -3781,13 +2781,27 @@ impl Panel for AgentPanel {
 
 impl AgentPanel {
     fn ensure_thread_initialized(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        if matches!(self.base_view, BaseView::Uninitialized)
-            && !matches!(
-                self.worktree_creation_status,
-                Some((_, WorktreeCreationStatus::Creating(_)))
-            )
-        {
+        if matches!(self.base_view, BaseView::Uninitialized) {
             self.activate_draft(false, window, cx);
+        }
+
+        self.apply_pending_worktree_draft(window, cx);
+    }
+
+    /// Tries to insert any existing draft prompt stashed in `pending_worktree_draft`
+    /// into the active thread view's message editor.
+    fn apply_pending_worktree_draft(&mut self, window: &mut Window, cx: &mut Context<Self>) {
+        if let Some(text) = self.pending_worktree_draft.take() {
+            if let Some(thread_view) = self.active_thread_view(cx) {
+                thread_view.update(cx, |thread_view, cx| {
+                    thread_view.message_editor.update(cx, |editor, cx| {
+                        editor.clear(window, cx);
+                        editor.insert_text(&text, window, cx);
+                    });
+                });
+            } else {
+                self.pending_worktree_draft = Some(text);
+            }
         }
     }
 
@@ -4052,10 +3066,6 @@ impl AgentPanel {
                     Tooltip::for_action_in("Go Back", &workspace::GoBack, &focus_handle, cx)
                 }
             })
-    }
-
-    fn project_has_git_repository(&self, cx: &App) -> bool {
-        !self.project.read(cx).repositories(cx).is_empty()
     }
 
     fn render_toolbar(&self, window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
@@ -4492,36 +3502,6 @@ impl AgentPanel {
             .child(toolbar_content)
     }
 
-    fn render_worktree_creation_status(&self, cx: &mut Context<Self>) -> Option<AnyElement> {
-        let (view_id, status) = self.worktree_creation_status.as_ref()?;
-        let active_view_id = self.active_conversation_view().map(|v| v.entity_id());
-        if active_view_id != Some(*view_id) {
-            return None;
-        }
-        match status {
-            WorktreeCreationStatus::Creating(_) | WorktreeCreationStatus::Loading(_) => None,
-            WorktreeCreationStatus::Error(message) => Some(
-                Callout::new()
-                    .icon(IconName::XCircleFilled)
-                    .severity(Severity::Error)
-                    .title("Worktree Creation Error")
-                    .description(message.clone())
-                    .border_position(ui::BorderPosition::Bottom)
-                    .dismiss_action(
-                        IconButton::new("dismiss-worktree-error", IconName::Close)
-                            .icon_size(IconSize::Small)
-                            .tooltip(Tooltip::text("Dismiss"))
-                            .on_click(cx.listener(|this, _, _, cx| {
-                                this.worktree_creation_status = None;
-                                cx.set_global(workspace::ActiveWorktreeCreation { label: None });
-                                cx.notify();
-                            })),
-                    )
-                    .into_any_element(),
-            ),
-        }
-    }
-
     fn should_render_trial_end_upsell(&self, cx: &mut Context<Self>) -> bool {
         if TrialEndUpsell::dismissed(cx) {
             return false;
@@ -4825,7 +3805,6 @@ impl Render for AgentPanel {
                     parent.children(configuration.cloned())
                 }
             })
-            .children(self.render_worktree_creation_status(cx))
             .children(self.render_trial_end_upsell(window, cx));
 
         match self.visible_font_size() {
@@ -4993,6 +3972,7 @@ impl AgentPanel {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::NewWorktreeBranchTarget;
     use crate::conversation_view::tests::{StubAgentServer, init_test};
     use crate::test_support::{
         active_session_id, active_thread_id, open_thread_with_connection,
@@ -5000,8 +3980,7 @@ mod tests {
     };
     use acp_thread::{AgentConnection, StubAgentConnection, ThreadStatus, UserMessageId};
     use action_log::ActionLog;
-    use agent_servers::CODEX_ID;
-    use anyhow::Result;
+    use anyhow::{Result, anyhow};
     use feature_flags::FeatureFlagAppExt;
     use fs::FakeFs;
     use gpui::{App, TestAppContext, VisualTestContext};
@@ -6272,339 +5251,39 @@ mod tests {
 
     #[gpui::test]
     fn test_resolve_worktree_branch_target() {
-        let resolved =
-            AgentPanel::resolve_worktree_branch_target(&NewWorktreeBranchTarget::CreateBranch {
+        let resolved = git_ui::worktree_service::resolve_worktree_branch_target(
+            &NewWorktreeBranchTarget::CreateBranch {
                 name: "new-branch".to_string(),
                 from_ref: Some("main".to_string()),
-            });
+            },
+        );
         assert_eq!(
             resolved,
             (Some("new-branch".to_string()), Some("main".to_string()))
         );
 
-        let resolved =
-            AgentPanel::resolve_worktree_branch_target(&NewWorktreeBranchTarget::CreateBranch {
+        let resolved = git_ui::worktree_service::resolve_worktree_branch_target(
+            &NewWorktreeBranchTarget::CreateBranch {
                 name: "new-branch".to_string(),
                 from_ref: None,
-            });
+            },
+        );
         assert_eq!(resolved, (Some("new-branch".to_string()), None));
 
-        let resolved =
-            AgentPanel::resolve_worktree_branch_target(&NewWorktreeBranchTarget::ExistingBranch {
+        let resolved = git_ui::worktree_service::resolve_worktree_branch_target(
+            &NewWorktreeBranchTarget::ExistingBranch {
                 name: "feature".to_string(),
-            });
+            },
+        );
         assert_eq!(
             resolved,
             (Some("feature".to_string()), Some("feature".to_string()))
         );
 
-        let resolved =
-            AgentPanel::resolve_worktree_branch_target(&NewWorktreeBranchTarget::CurrentBranch);
+        let resolved = git_ui::worktree_service::resolve_worktree_branch_target(
+            &NewWorktreeBranchTarget::CurrentBranch,
+        );
         assert_eq!(resolved, (None, None));
-    }
-
-    #[gpui::test]
-    async fn test_worktree_dir_name_is_random_when_using_existing_branch(cx: &mut TestAppContext) {
-        init_test(cx);
-
-        let app_state = cx.update(|cx| {
-            agent::ThreadStore::init_global(cx);
-            language_model::LanguageModelRegistry::test(cx);
-
-            let app_state = workspace::AppState::test(cx);
-            workspace::init(app_state.clone(), cx);
-            app_state
-        });
-
-        let fs = app_state.fs.as_fake();
-        fs.insert_tree(
-            "/project",
-            json!({
-                ".git": {},
-                "src": {
-                    "main.rs": "fn main() {}"
-                }
-            }),
-        )
-        .await;
-        // Put the main worktree on "develop" so that "main" is NOT
-        // occupied by any worktree.
-        fs.set_branch_name(Path::new("/project/.git"), Some("develop"));
-        fs.insert_branches(Path::new("/project/.git"), &["main", "develop"]);
-
-        let project = Project::test(app_state.fs.clone(), [Path::new("/project")], cx).await;
-
-        let multi_workspace =
-            cx.add_window(|window, cx| MultiWorkspace::test_new(project.clone(), window, cx));
-        multi_workspace
-            .update(cx, |multi_workspace, _, cx| {
-                multi_workspace.open_sidebar(cx);
-            })
-            .unwrap();
-
-        let workspace = multi_workspace
-            .read_with(cx, |multi_workspace, _cx| {
-                multi_workspace.workspace().clone()
-            })
-            .unwrap();
-
-        workspace.update(cx, |workspace, _cx| {
-            workspace.set_random_database_id();
-        });
-
-        cx.update(|cx| {
-            cx.observe_new(
-                |workspace: &mut Workspace,
-                 window: Option<&mut Window>,
-                 cx: &mut Context<Workspace>| {
-                    if let Some(window) = window {
-                        let panel = cx.new(|cx| AgentPanel::new(workspace, None, window, cx));
-                        workspace.add_panel(panel, window, cx);
-                    }
-                },
-            )
-            .detach();
-        });
-
-        let cx = &mut VisualTestContext::from_window(multi_workspace.into(), cx);
-        cx.run_until_parked();
-
-        let panel = workspace.update_in(cx, |workspace, window, cx| {
-            let panel = cx.new(|cx| AgentPanel::new(workspace, None, window, cx));
-            workspace.add_panel(panel.clone(), window, cx);
-            panel
-        });
-
-        cx.run_until_parked();
-
-        panel.update_in(cx, |panel, window, cx| {
-            panel.open_external_thread_with_server(
-                Rc::new(StubAgentServer::default_response()),
-                window,
-                cx,
-            );
-        });
-
-        cx.run_until_parked();
-
-        // Select "main" as an existing branch — this should NOT make the
-        // worktree directory named "main"; it should get a random name.
-        let content = vec![acp::ContentBlock::Text(acp::TextContent::new(
-            "Hello from test",
-        ))];
-        panel.update_in(cx, |panel, window, cx| {
-            panel.handle_worktree_requested(
-                content,
-                WorktreeCreationArgs::New {
-                    worktree_name: None,
-                    branch_target: NewWorktreeBranchTarget::ExistingBranch {
-                        name: "main".to_string(),
-                    },
-                },
-                PreviousWorkspaceState::empty(),
-                window,
-                cx,
-            );
-        });
-
-        cx.run_until_parked();
-
-        // Find the new workspace and check its worktree path.
-        let new_worktree_path = multi_workspace
-            .read_with(cx, |multi_workspace, cx| {
-                let new_workspace = multi_workspace
-                    .workspaces()
-                    .find(|ws| ws.entity_id() != workspace.entity_id())
-                    .expect("a new workspace should have been created");
-
-                let new_project = new_workspace.read(cx).project().clone();
-                let worktree = new_project
-                    .read(cx)
-                    .visible_worktrees(cx)
-                    .next()
-                    .expect("new workspace should have a worktree");
-                worktree.read(cx).abs_path().to_path_buf()
-            })
-            .unwrap();
-
-        // The worktree directory path should contain a random adjective-noun
-        // name, NOT the branch name "main".
-        let path_str = new_worktree_path.to_string_lossy();
-        assert!(
-            !path_str.contains("/main/"),
-            "worktree directory should use a random name, not the branch name. \
-             Got path: {path_str}",
-        );
-        // Verify it looks like an adjective-noun pair (contains a hyphen in
-        // the directory component above the project name).
-        let parent = new_worktree_path
-            .parent()
-            .and_then(|p| p.file_name())
-            .and_then(|n| n.to_str())
-            .expect("should have a parent directory name");
-        assert!(
-            parent.contains('-'),
-            "worktree parent directory should be an adjective-noun pair (e.g. 'swift-falcon'), \
-             got: {parent}",
-        );
-    }
-
-    #[gpui::test]
-    async fn test_worktree_creation_preserves_selected_agent(cx: &mut TestAppContext) {
-        init_test(cx);
-
-        let app_state = cx.update(|cx| {
-            agent::ThreadStore::init_global(cx);
-            language_model::LanguageModelRegistry::test(cx);
-
-            let app_state = workspace::AppState::test(cx);
-            workspace::init(app_state.clone(), cx);
-            app_state
-        });
-
-        let fs = app_state.fs.as_fake();
-        fs.insert_tree(
-            "/project",
-            json!({
-                ".git": {},
-                "src": {
-                    "main.rs": "fn main() {}"
-                }
-            }),
-        )
-        .await;
-        fs.set_branch_name(Path::new("/project/.git"), Some("main"));
-
-        let project = Project::test(app_state.fs.clone(), [Path::new("/project")], cx).await;
-
-        let multi_workspace =
-            cx.add_window(|window, cx| MultiWorkspace::test_new(project.clone(), window, cx));
-        multi_workspace
-            .update(cx, |multi_workspace, _, cx| {
-                multi_workspace.open_sidebar(cx);
-            })
-            .unwrap();
-
-        let workspace = multi_workspace
-            .read_with(cx, |multi_workspace, _cx| {
-                multi_workspace.workspace().clone()
-            })
-            .unwrap();
-
-        workspace.update(cx, |workspace, _cx| {
-            workspace.set_random_database_id();
-        });
-
-        // Register a callback so new workspaces also get an AgentPanel.
-        cx.update(|cx| {
-            cx.observe_new(
-                |workspace: &mut Workspace,
-                 window: Option<&mut Window>,
-                 cx: &mut Context<Workspace>| {
-                    if let Some(window) = window {
-                        let panel = cx.new(|cx| AgentPanel::new(workspace, None, window, cx));
-                        workspace.add_panel(panel, window, cx);
-                    }
-                },
-            )
-            .detach();
-        });
-
-        let cx = &mut VisualTestContext::from_window(multi_workspace.into(), cx);
-
-        // Wait for the project to discover the git repository.
-        cx.run_until_parked();
-
-        let panel = workspace.update_in(cx, |workspace, window, cx| {
-            let panel = cx.new(|cx| AgentPanel::new(workspace, None, window, cx));
-            workspace.add_panel(panel.clone(), window, cx);
-            panel
-        });
-
-        cx.run_until_parked();
-
-        // Open a thread (needed so there's an active thread view).
-        panel.update_in(cx, |panel, window, cx| {
-            panel.open_external_thread_with_server(
-                Rc::new(StubAgentServer::default_response()),
-                window,
-                cx,
-            );
-        });
-
-        cx.run_until_parked();
-
-        // Set the selected agent to Codex (a custom agent). We do this AFTER
-        // opening the thread because open_external_thread_with_server overrides
-        // selected_agent.
-        panel.update_in(cx, |panel, _window, cx| {
-            panel.selected_agent = Agent::Custom {
-                id: CODEX_ID.into(),
-            };
-            cx.notify();
-        });
-
-        // Verify the panel has the Codex agent selected.
-        panel.read_with(cx, |panel, _cx| {
-            assert_eq!(
-                panel.selected_agent,
-                Agent::Custom {
-                    id: CODEX_ID.into()
-                },
-            );
-        });
-
-        // Directly call handle_worktree_requested to trigger worktree creation.
-        let content = vec![acp::ContentBlock::Text(acp::TextContent::new(
-            "Hello from test",
-        ))];
-        panel.update_in(cx, |panel, window, cx| {
-            panel.handle_worktree_requested(
-                content,
-                WorktreeCreationArgs::New {
-                    worktree_name: None,
-                    branch_target: NewWorktreeBranchTarget::default(),
-                },
-                PreviousWorkspaceState::empty(),
-                window,
-                cx,
-            );
-        });
-
-        // Let the async worktree creation + workspace setup complete.
-        cx.run_until_parked();
-
-        // Find the new workspace's AgentPanel and verify it used the Codex agent.
-        let found_codex = multi_workspace
-            .read_with(cx, |multi_workspace, cx| {
-                // There should be more than one workspace now (the original + the new worktree).
-                assert!(
-                    multi_workspace.workspaces().count() > 1,
-                    "expected a new workspace to have been created, found {}",
-                    multi_workspace.workspaces().count(),
-                );
-
-                // Check the newest workspace's panel for the correct agent.
-                let new_workspace = multi_workspace
-                    .workspaces()
-                    .find(|ws| ws.entity_id() != workspace.entity_id())
-                    .expect("should find the new workspace");
-                let new_panel = new_workspace
-                    .read(cx)
-                    .panel::<AgentPanel>(cx)
-                    .expect("new workspace should have an AgentPanel");
-
-                new_panel.read(cx).selected_agent.clone()
-            })
-            .unwrap();
-
-        assert_eq!(
-            found_codex,
-            Agent::Custom {
-                id: CODEX_ID.into()
-            },
-            "the new worktree workspace should use the same agent (Codex) that was selected in the original panel",
-        );
     }
 
     #[gpui::test]
@@ -7142,7 +5821,12 @@ mod tests {
         let result = multi_workspace
             .update(cx, |_, window, cx| {
                 window.spawn(cx, async move |cx| {
-                    AgentPanel::await_and_rollback_on_failure(creation_infos, fs_clone, cx).await
+                    git_ui::worktree_service::await_and_rollback_on_failure(
+                        creation_infos,
+                        fs_clone,
+                        cx,
+                    )
+                    .await
                 })
             })
             .unwrap()
@@ -7225,7 +5909,12 @@ mod tests {
         let result = multi_workspace
             .update(cx, |_, window, cx| {
                 window.spawn(cx, async move |cx| {
-                    AgentPanel::await_and_rollback_on_failure(creation_infos, fs_clone, cx).await
+                    git_ui::worktree_service::await_and_rollback_on_failure(
+                        creation_infos,
+                        fs_clone,
+                        cx,
+                    )
+                    .await
                 })
             })
             .unwrap()
@@ -7291,7 +5980,12 @@ mod tests {
         let result = multi_workspace
             .update(cx, |_, window, cx| {
                 window.spawn(cx, async move |cx| {
-                    AgentPanel::await_and_rollback_on_failure(creation_infos, fs_clone, cx).await
+                    git_ui::worktree_service::await_and_rollback_on_failure(
+                        creation_infos,
+                        fs_clone,
+                        cx,
+                    )
+                    .await
                 })
             })
             .unwrap()
@@ -7361,7 +6055,12 @@ mod tests {
         let result = multi_workspace
             .update(cx, |_, window, cx| {
                 window.spawn(cx, async move |cx| {
-                    AgentPanel::await_and_rollback_on_failure(creation_infos, fs_clone, cx).await
+                    git_ui::worktree_service::await_and_rollback_on_failure(
+                        creation_infos,
+                        fs_clone,
+                        cx,
+                    )
+                    .await
                 })
             })
             .unwrap()
@@ -7373,409 +6072,6 @@ mod tests {
         assert!(
             !fs.is_dir(&orphan_path).await,
             "orphan worktree directory should be removed by filesystem cleanup"
-        );
-    }
-
-    #[gpui::test]
-    async fn test_worktree_creation_for_remote_project(
-        cx: &mut TestAppContext,
-        server_cx: &mut TestAppContext,
-    ) {
-        init_test(cx);
-
-        let app_state = cx.update(|cx| {
-            agent::ThreadStore::init_global(cx);
-            language_model::LanguageModelRegistry::test(cx);
-
-            let app_state = workspace::AppState::test(cx);
-            workspace::init(app_state.clone(), cx);
-            app_state
-        });
-
-        server_cx.update(|cx| {
-            release_channel::init(semver::Version::new(0, 0, 0), cx);
-        });
-
-        // Set up the remote server side with a git repo.
-        let server_fs = FakeFs::new(server_cx.executor());
-        server_fs
-            .insert_tree(
-                "/project",
-                json!({
-                    ".git": {},
-                    "src": {
-                        "main.rs": "fn main() {}"
-                    }
-                }),
-            )
-            .await;
-        server_fs.set_branch_name(Path::new("/project/.git"), Some("main"));
-
-        // Create a mock remote connection.
-        let (opts, server_session, _) = remote::RemoteClient::fake_server(cx, server_cx);
-
-        server_cx.update(remote_server::HeadlessProject::init);
-        let server_executor = server_cx.executor();
-        let _headless = server_cx.new(|cx| {
-            remote_server::HeadlessProject::new(
-                remote_server::HeadlessAppState {
-                    session: server_session,
-                    fs: server_fs.clone(),
-                    http_client: Arc::new(http_client::BlockedHttpClient),
-                    node_runtime: node_runtime::NodeRuntime::unavailable(),
-                    languages: Arc::new(language::LanguageRegistry::new(server_executor.clone())),
-                    extension_host_proxy: Arc::new(extension::ExtensionHostProxy::new()),
-                    startup_time: Instant::now(),
-                },
-                false,
-                cx,
-            )
-        });
-
-        // Connect the client side and build a remote project.
-        // Use a separate Client to avoid double-registering proto handlers
-        // (Workspace::test_new creates its own WorkspaceStore from the
-        // project's client).
-        let remote_client = remote::RemoteClient::connect_mock(opts, cx).await;
-        let project = cx.update(|cx| {
-            let project_client = client::Client::new(
-                Arc::new(clock::FakeSystemClock::new()),
-                http_client::FakeHttpClient::with_404_response(),
-                cx,
-            );
-            let user_store = cx.new(|cx| client::UserStore::new(project_client.clone(), cx));
-            project::Project::remote(
-                remote_client,
-                project_client,
-                node_runtime::NodeRuntime::unavailable(),
-                user_store,
-                app_state.languages.clone(),
-                app_state.fs.clone(),
-                false,
-                cx,
-            )
-        });
-
-        // Open the remote path as a worktree in the project.
-        let worktree_path = Path::new("/project");
-        project
-            .update(cx, |project, cx| {
-                project.find_or_create_worktree(worktree_path, true, cx)
-            })
-            .await
-            .expect("should be able to open remote worktree");
-        cx.run_until_parked();
-
-        // Verify the project is indeed remote.
-        project.read_with(cx, |project, cx| {
-            assert!(!project.is_local(), "project should be remote, not local");
-            assert!(
-                project.remote_connection_options(cx).is_some(),
-                "project should have remote connection options"
-            );
-        });
-
-        // Create the workspace and agent panel.
-        let multi_workspace =
-            cx.add_window(|window, cx| MultiWorkspace::test_new(project.clone(), window, cx));
-        multi_workspace
-            .update(cx, |multi_workspace, _, cx| {
-                multi_workspace.open_sidebar(cx);
-            })
-            .unwrap();
-
-        let workspace = multi_workspace
-            .read_with(cx, |mw, _cx| mw.workspace().clone())
-            .unwrap();
-
-        workspace.update(cx, |workspace, _cx| {
-            workspace.set_random_database_id();
-        });
-
-        // Register a callback so new workspaces also get an AgentPanel.
-        cx.update(|cx| {
-            cx.observe_new(
-                |workspace: &mut Workspace,
-                 window: Option<&mut Window>,
-                 cx: &mut Context<Workspace>| {
-                    if let Some(window) = window {
-                        let panel = cx.new(|cx| AgentPanel::new(workspace, None, window, cx));
-                        workspace.add_panel(panel, window, cx);
-                    }
-                },
-            )
-            .detach();
-        });
-
-        let cx = &mut VisualTestContext::from_window(multi_workspace.into(), cx);
-        cx.run_until_parked();
-
-        let panel = workspace.update_in(cx, |workspace, window, cx| {
-            let panel = cx.new(|cx| AgentPanel::new(workspace, None, window, cx));
-            workspace.add_panel(panel.clone(), window, cx);
-            panel
-        });
-
-        cx.run_until_parked();
-
-        // Open a thread.
-        panel.update_in(cx, |panel, window, cx| {
-            panel.open_external_thread_with_server(
-                Rc::new(StubAgentServer::default_response()),
-                window,
-                cx,
-            );
-        });
-        cx.run_until_parked();
-
-        // Trigger worktree creation for a known linked path.
-        let linked_path = PathBuf::from("/project");
-        let content = vec![acp::ContentBlock::Text(acp::TextContent::new(
-            "Hello from remote test",
-        ))];
-        panel.update_in(cx, |panel, window, cx| {
-            panel.handle_worktree_requested(
-                content,
-                WorktreeCreationArgs::Linked {
-                    worktree_path: linked_path,
-                    display_name: "test-worktree".to_string(),
-                },
-                PreviousWorkspaceState::empty(),
-                window,
-                cx,
-            );
-        });
-
-        // The refactored code uses `find_or_create_workspace`, which
-        // finds the existing remote workspace (matching paths + host)
-        // and reuses it instead of creating a new connection.
-        cx.run_until_parked();
-
-        // The task should have completed: the existing workspace was
-        // found and reused.
-        panel.read_with(cx, |panel, _cx| {
-            assert!(
-                panel.worktree_creation_status.is_none(),
-                "worktree creation should have completed, but status is: {:?}",
-                panel.worktree_creation_status
-            );
-        });
-
-        // The existing remote workspace was reused — no new workspace
-        // should have been created.
-        multi_workspace
-            .read_with(cx, |multi_workspace, cx| {
-                let project = workspace.read(cx).project().clone();
-                assert!(
-                    !project.read(cx).is_local(),
-                    "workspace project should still be remote, not local"
-                );
-                assert_eq!(
-                    multi_workspace.workspaces().count(),
-                    1,
-                    "existing remote workspace should be reused, not a new one created"
-                );
-            })
-            .unwrap();
-    }
-
-    #[gpui::test]
-    async fn test_linked_worktree_switch_remaps_open_files(cx: &mut TestAppContext) {
-        init_test(cx);
-
-        let app_state = cx.update(|cx| {
-            agent::ThreadStore::init_global(cx);
-            language_model::LanguageModelRegistry::test(cx);
-
-            let app_state = workspace::AppState::test(cx);
-            workspace::init(app_state.clone(), cx);
-            app_state
-        });
-
-        let fs = app_state.fs.as_fake();
-        fs.insert_tree(
-            "/project",
-            json!({
-                ".git": {},
-                "src": {
-                    "main.rs": "fn main() {}",
-                    "lib.rs": "pub fn hello() {}"
-                }
-            }),
-        )
-        .await;
-        fs.set_branch_name(Path::new("/project/.git"), Some("main"));
-
-        // Create a linked worktree directory with the same file structure.
-        let linked_path = PathBuf::from("/linked-worktree");
-        fs.add_linked_worktree_for_repo(
-            Path::new("/project/.git"),
-            true,
-            git::repository::Worktree {
-                path: linked_path.clone(),
-                ref_name: Some("refs/heads/feature".into()),
-                sha: "abc123".into(),
-                is_main: false,
-                is_bare: false,
-            },
-        )
-        .await;
-        fs.insert_tree(
-            "/linked-worktree",
-            json!({
-                "src": {
-                    "main.rs": "fn main() { // linked }",
-                    "lib.rs": "pub fn hello() { // linked }"
-                }
-            }),
-        )
-        .await;
-
-        let project = Project::test(app_state.fs.clone(), [Path::new("/project")], cx).await;
-
-        let multi_workspace =
-            cx.add_window(|window, cx| MultiWorkspace::test_new(project.clone(), window, cx));
-        multi_workspace
-            .update(cx, |multi_workspace, _, cx| {
-                multi_workspace.open_sidebar(cx);
-            })
-            .unwrap();
-
-        let workspace = multi_workspace
-            .read_with(cx, |multi_workspace, _cx| {
-                multi_workspace.workspace().clone()
-            })
-            .unwrap();
-
-        workspace.update(cx, |workspace, _cx| {
-            workspace.set_random_database_id();
-        });
-
-        // Register observer so new workspaces get AgentPanel automatically.
-        cx.update(|cx| {
-            cx.observe_new(
-                |workspace: &mut Workspace,
-                 window: Option<&mut Window>,
-                 cx: &mut Context<Workspace>| {
-                    if let Some(window) = window {
-                        let panel = cx.new(|cx| AgentPanel::new(workspace, None, window, cx));
-                        workspace.add_panel(panel, window, cx);
-                    }
-                },
-            )
-            .detach();
-        });
-
-        let cx = &mut VisualTestContext::from_window(multi_workspace.into(), cx);
-        cx.run_until_parked();
-
-        let panel = workspace.update_in(cx, |workspace, window, cx| {
-            let panel = cx.new(|cx| AgentPanel::new(workspace, None, window, cx));
-            workspace.add_panel(panel.clone(), window, cx);
-            panel
-        });
-
-        cx.run_until_parked();
-
-        // Open files in the original workspace.
-        workspace
-            .update_in(cx, |workspace, window, cx| {
-                workspace.open_paths(
-                    vec![
-                        PathBuf::from("/project/src/main.rs"),
-                        PathBuf::from("/project/src/lib.rs"),
-                    ],
-                    workspace::OpenOptions::default(),
-                    None,
-                    window,
-                    cx,
-                )
-            })
-            .await;
-        cx.run_until_parked();
-
-        // Verify files are open.
-        workspace.read_with(cx, |workspace, cx| {
-            let open_paths = workspace.open_item_abs_paths(cx);
-            assert!(
-                open_paths.iter().any(|p| p.ends_with("src/main.rs")),
-                "main.rs should be open, got: {open_paths:?}"
-            );
-            assert!(
-                open_paths.iter().any(|p| p.ends_with("src/lib.rs")),
-                "lib.rs should be open, got: {open_paths:?}"
-            );
-        });
-
-        // Open a thread so the panel is in a valid state.
-        panel.update_in(cx, |panel, window, cx| {
-            panel.open_external_thread_with_server(
-                Rc::new(StubAgentServer::default_response()),
-                window,
-                cx,
-            );
-        });
-        cx.run_until_parked();
-
-        // Build a PreviousWorkspaceState with the open files.
-        let previous_state =
-            workspace.update_in(cx, |workspace, window, cx| PreviousWorkspaceState {
-                dock_structure: workspace.capture_dock_state(window, cx),
-                open_file_paths: vec![
-                    PathBuf::from("/project/src/main.rs"),
-                    PathBuf::from("/project/src/lib.rs"),
-                ],
-                active_file_path: Some(PathBuf::from("/project/src/main.rs")),
-            });
-
-        // Trigger the linked worktree switch.
-        let content = vec![acp::ContentBlock::Text(acp::TextContent::new(
-            "Hello from linked worktree test",
-        ))];
-        panel.update_in(cx, |panel, window, cx| {
-            panel.handle_worktree_requested(
-                content,
-                WorktreeCreationArgs::Linked {
-                    worktree_path: linked_path.clone(),
-                    display_name: "feature".to_string(),
-                },
-                previous_state,
-                window,
-                cx,
-            );
-        });
-
-        cx.run_until_parked();
-
-        // Find the new workspace.
-        let new_workspace = multi_workspace
-            .read_with(cx, |multi_workspace, _cx| {
-                multi_workspace
-                    .workspaces()
-                    .find(|ws| ws.entity_id() != workspace.entity_id())
-                    .cloned()
-            })
-            .unwrap()
-            .expect("a new workspace should have been created for the linked worktree");
-
-        // Verify that files were remapped and opened in the new workspace.
-        // The original /project/src/main.rs should now be /linked-worktree/src/main.rs.
-        let new_open_paths =
-            new_workspace.read_with(cx, |workspace, cx| workspace.open_item_abs_paths(cx));
-
-        assert!(
-            new_open_paths
-                .iter()
-                .any(|p| p == &linked_path.join("src/main.rs")),
-            "main.rs should have been remapped to the linked worktree. \
-             Open paths: {new_open_paths:?}"
-        );
-        assert!(
-            new_open_paths
-                .iter()
-                .any(|p| p == &linked_path.join("src/lib.rs")),
-            "lib.rs should have been remapped to the linked worktree. \
-             Open paths: {new_open_paths:?}"
         );
     }
 
@@ -7909,7 +6205,8 @@ mod tests {
         cx.run_until_parked();
 
         panel.read_with(cx, |panel, cx| {
-            let (git_repos, non_git_paths) = panel.classify_worktrees(cx);
+            let (git_repos, non_git_paths) =
+                git_ui::worktree_service::classify_worktrees(panel.project.read(cx), cx);
 
             let git_work_dirs: Vec<PathBuf> = git_repos
                 .iter()
