+++ crates/git_ui/Cargo.toml
@@ -27,6 +27,7 @@ component.workspace = true
 db.workspace = true
 editor.workspace = true
 file_icons.workspace = true
+fs.workspace = true
 futures.workspace = true
 fuzzy.workspace = true
 git.workspace = true
@@ -45,6 +46,7 @@ picker.workspace = true
 project.workspace = true
 prompt_store.workspace = true
 proto.workspace = true
+rand.workspace = true
 remote_connection.workspace = true
 remote.workspace = true
 schemars.workspace = true
