+++ crates/git_ui/src/worktree_picker.rs
@@ -41,6 +41,12 @@ impl WorktreePicker {
         window: &mut Window,
         cx: &mut Context<Self>,
     ) -> Self {
+        if let Some(workspace) = workspace.upgrade() {
+            let focused_dock = workspace.read(cx).focused_dock_position(window, cx);
+            workspace.update(cx, |workspace, _cx| {
+                workspace.set_pre_picker_focused_dock(focused_dock);
+            });
+        }
         Self::new_inner(project, workspace, false, window, cx)
     }
 
@@ -83,11 +89,11 @@ impl WorktreePicker {
 
         let all_worktrees_request = repository
             .clone()
-            .map(|repo| repo.update(cx, |repo, _| repo.worktrees()));
+            .map(|repository| repository.update(cx, |repository, _| repository.worktrees()));
 
-        let default_branch_request = repository
-            .clone()
-            .map(|repo| repo.update(cx, |repo, _| repo.default_branch(false)));
+        let default_branch_request = repository.clone().map(|repository| {
+            repository.update(cx, |repository, _| repository.default_branch(false))
+        });
 
         let initial_matches = vec![WorktreeEntry::CreateFromCurrentBranch];
 
@@ -108,6 +114,7 @@ impl WorktreePicker {
         let picker = cx.new(|cx| {
             Picker::list(delegate, window, cx)
                 .list_measure_all()
+                .show_scrollbar(true)
                 .modal(false)
                 .max_height(Some(rems(20.).into()))
         });
@@ -550,7 +557,6 @@ impl PickerDelegate for WorktreePickerDelegate {
 
         match entry {
             WorktreeEntry::Separator => return,
-
             WorktreeEntry::CreateFromCurrentBranch => {
                 window.dispatch_action(
                     Box::new(CreateWorktree {
@@ -560,7 +566,6 @@ impl PickerDelegate for WorktreePickerDelegate {
                     cx,
                 );
             }
-
             WorktreeEntry::CreateFromDefaultBranch {
                 default_branch_name,
             } => {
@@ -574,7 +579,6 @@ impl PickerDelegate for WorktreePickerDelegate {
                     cx,
                 );
             }
-
             WorktreeEntry::Worktree { worktree, .. } => {
                 let is_current = self.project_worktree_paths.contains(&worktree.path);
 
@@ -602,7 +606,6 @@ impl PickerDelegate for WorktreePickerDelegate {
                     }
                 }
             }
-
             WorktreeEntry::CreateNamed {
                 name,
                 from_branch,
@@ -622,7 +625,6 @@ impl PickerDelegate for WorktreePickerDelegate {
                     cx,
                 );
             }
-
             WorktreeEntry::CreateNamed {
                 disabled_reason: Some(_),
                 ..
@@ -649,40 +651,6 @@ impl PickerDelegate for WorktreePickerDelegate {
 
         let no_git_reason: SharedString = "Requires a Git repository in the project".into();
 
-        let create_new_list_item = |id: SharedString,
-                                    label: SharedString,
-                                    disabled_tooltip: Option<SharedString>,
-                                    selected: bool| {
-            let is_disabled = disabled_tooltip.is_some();
-            ListItem::new(id)
-                .inset(true)
-                .spacing(ListItemSpacing::Sparse)
-                .toggle_state(selected)
-                .child(
-                    h_flex()
-                        .w_full()
-                        .gap_2p5()
-                        .child(
-                            Icon::new(IconName::Plus)
-                                .map(|this| {
-                                    if is_disabled {
-                                        this.color(Color::Disabled)
-                                    } else {
-                                        this.color(Color::Muted)
-                                    }
-                                })
-                                .size(IconSize::Small),
-                        )
-                        .child(
-                            Label::new(label).when(is_disabled, |this| this.color(Color::Disabled)),
-                        ),
-                )
-                .when_some(disabled_tooltip, |this, reason| {
-                    this.tooltip(Tooltip::text(reason))
-                })
-                .into_any_element()
-        };
-
         match entry {
             WorktreeEntry::Separator => Some(
                 div()
@@ -690,7 +658,6 @@ impl PickerDelegate for WorktreePickerDelegate {
                     .child(Divider::horizontal())
                     .into_any_element(),
             ),
-
             WorktreeEntry::CreateFromCurrentBranch => {
                 let branch_label = if self.has_multiple_repositories {
                     "current branches".to_string()
@@ -713,7 +680,6 @@ impl PickerDelegate for WorktreePickerDelegate {
 
                 Some(item.into_any_element())
             }
-
             WorktreeEntry::CreateFromDefaultBranch {
                 default_branch_name,
             } => {
@@ -730,7 +696,6 @@ impl PickerDelegate for WorktreePickerDelegate {
 
                 Some(item.into_any_element())
             }
-
             WorktreeEntry::Worktree {
                 worktree,
                 positions,
@@ -876,7 +841,6 @@ impl PickerDelegate for WorktreePickerDelegate {
                         .into_any_element(),
                 )
             }
-
             WorktreeEntry::CreateNamed {
                 name,
                 from_branch,
@@ -1003,6 +967,41 @@ impl PickerDelegate for WorktreePickerDelegate {
             None
         }
     }
+}
+
+fn create_new_list_item(
+    id: SharedString,
+    label: SharedString,
+    disabled_tooltip: Option<SharedString>,
+    selected: bool,
+) -> AnyElement {
+    let is_disabled = disabled_tooltip.is_some();
+
+    ListItem::new(id)
+        .inset(true)
+        .spacing(ListItemSpacing::Sparse)
+        .toggle_state(selected)
+        .child(
+            h_flex()
+                .w_full()
+                .gap_2p5()
+                .child(
+                    Icon::new(IconName::Plus)
+                        .map(|this| {
+                            if is_disabled {
+                                this.color(Color::Disabled)
+                            } else {
+                                this.color(Color::Muted)
+                            }
+                        })
+                        .size(IconSize::Small),
+                )
+                .child(Label::new(label).when(is_disabled, |this| this.color(Color::Disabled))),
+        )
+        .when_some(disabled_tooltip, |this, reason| {
+            this.tooltip(Tooltip::text(reason))
+        })
+        .into_any_element()
 }
 
 pub async fn open_remote_worktree(
