+++ crates/migrator/src/migrations.rs
@@ -334,3 +334,9 @@ pub(crate) mod m_2026_04_10 {
 
     pub(crate) use settings::rename_web_search_to_search_web;
 }
+
+pub(crate) mod m_2026_04_17 {
+    mod settings;
+
+    pub(crate) use settings::SETTINGS_PATTERNS;
+}
