+++ crates/migrator/src/migrator.rs
@@ -250,6 +250,10 @@ pub fn migrate_settings(text: &str) -> Result<Option<String>> {
         MigrationType::Json(migrations::m_2026_03_30::make_play_sound_when_agent_done_an_enum),
         MigrationType::Json(migrations::m_2026_04_01::restructure_profiles_with_settings_key),
         MigrationType::Json(migrations::m_2026_04_10::rename_web_search_to_search_web),
+        MigrationType::TreeSitter(
+            migrations::m_2026_04_17::SETTINGS_PATTERNS,
+            &SETTINGS_QUERY_2026_04_17,
+        ),
     ];
     run_migrations(text, migrations)
 }
@@ -391,6 +395,10 @@ define_query!(
 define_query!(
     KEYMAP_QUERY_2026_03_23,
     migrations::m_2026_03_23::KEYMAP_PATTERNS
+);
+define_query!(
+    SETTINGS_QUERY_2026_04_17,
+    migrations::m_2026_04_17::SETTINGS_PATTERNS
 );
 
 // custom query
@@ -5024,6 +5032,48 @@ mod tests {
     }
 }"#,
             ),
+        );
+    }
+
+    #[test]
+    fn test_rename_show_branch_icon_to_show_branch_status_icon() {
+        assert_migrate_settings(
+            &r#"
+            {
+                "title_bar": {
+                    "show_branch_icon": true,
+                    "show_branch_name": true
+                }
+            }
+            "#
+            .unindent(),
+            Some(
+                &r#"
+                {
+                    "title_bar": {
+                        "show_branch_status_icon": true,
+                        "show_branch_name": true
+                    }
+                }
+                "#
+                .unindent(),
+            ),
+        );
+
+        // No title_bar key — should be unchanged
+        assert_migrate_settings(&r#"{ "theme": "One Dark" }"#.unindent(), None);
+
+        // title_bar without show_branch_icon — should be unchanged
+        assert_migrate_settings(
+            &r#"
+            {
+                "title_bar": {
+                    "show_branch_name": true
+                }
+            }
+            "#
+            .unindent(),
+            None,
         );
     }
 }
