+++ crates/settings_content/src/title_bar.rs
@@ -81,10 +81,12 @@ impl From<String> for WindowButtonLayoutContent {
 #[with_fallible_options]
 #[derive(Clone, PartialEq, Default, Serialize, Deserialize, JsonSchema, MergeFrom, Debug)]
 pub struct TitleBarSettingsContent {
-    /// Whether to show the branch icon beside branch switcher in the title bar.
+    /// Whether to show git status indicators on the branch icon in the title bar.
+    /// When enabled, the branch icon changes to reflect the current repository
+    /// status (e.g. modified, added, deleted, or conflict).
     ///
     /// Default: false
-    pub show_branch_icon: Option<bool>,
+    pub show_branch_status_icon: Option<bool>,
     /// Whether to show onboarding banners in the title bar.
     ///
     /// Default: true
